refactor(benchmark): Use kebab-case for file and directory names
