+++ packages/@n8n/benchmark/scenarios/binary-data/binary-data.manifest.json
@@ -2,6 +2,6 @@
 	"$schema": "../scenario.schema.json",
 	"name": "BinaryData",
 	"description": "Send a binary file to a webhook, write it to FS, read it from FS and receive it back",
-	"scenarioData": { "workflowFiles": ["binaryData.json"] },
-	"scriptPath": "binaryData.script.js"
+	"scenarioData": { "workflowFiles": ["binary-data.json"] },
+	"scriptPath": "binary-data.script.js"
 }
