+++ packages/@n8n/benchmark/scenarios/binary-data/binary-data.script.js
