+++ packages/@n8n/benchmark/scenarios/http-node/http-node.json
