+++ packages/@n8n/benchmark/scenarios/http-node/http-node.manifest.json
@@ -2,6 +2,6 @@
 	"$schema": "../scenario.schema.json",
 	"name": "HttpNode",
 	"description": "Webhook -> 3x HTTP request to a mock API -> Merge -> Respond to Webhook. Requires a mock API running at http://mockapi:8080",
-	"scenarioData": { "workflowFiles": ["httpNode.json"] },
-	"scriptPath": "httpNode.script.js"
+	"scenarioData": { "workflowFiles": ["http-node.json"] },
+	"scriptPath": "http-node.script.js"
 }
