+++ packages/@n8n/benchmark/scenarios/js-code-node-once-for-each/js-code-node-once-for-each.manifest.json
@@ -2,6 +2,6 @@
 	"$schema": "../scenario.schema.json",
 	"name": "CodeNodeJsOnceForEach",
 	"description": "A JS Code Node that runs once for each item and adds, modifies and removes properties. The data of 5 items is generated using DebugHelper Node, and returned with RespondToWebhook Node.",
-	"scenarioData": { "workflowFiles": ["jsCodeNodeOnceForEach.json"] },
-	"scriptPath": "jsCodeNodeOnceForEach.script.js"
+	"scenarioData": { "workflowFiles": ["js-code-node-once-for-each.json"] },
+	"scriptPath": "js-code-node-once-for-each.script.js"
 }
