+++ packages/@n8n/benchmark/scenarios/js-code-node-once-for-each/js-code-node-once-for-each.script.js
