+++ packages/@n8n/benchmark/scenarios/set-node-expressions/set-node-expressions.json
