+++ packages/@n8n/benchmark/scenarios/single-webhook/single-webhook.json
