+++ packages/@n8n/benchmark/scenarios/single-webhook/single-webhook.manifest.json
@@ -2,6 +2,6 @@
 	"$schema": "../scenario.schema.json",
 	"name": "SingleWebhook",
 	"description": "A single webhook trigger that responds with a 200 status code",
-	"scenarioData": { "workflowFiles": ["singleWebhook.json"] },
-	"scriptPath": "singleWebhook.script.ts"
+	"scenarioData": { "workflowFiles": ["single-webhook.json"] },
+	"scriptPath": "single-webhook.script.ts"
 }
