+++ packages/@n8n/benchmark/scripts/clients/docker-compose-client.mjs
