+++ packages/@n8n/benchmark/scripts/clients/ssh-client.mjs
