+++ packages/@n8n/benchmark/scripts/n8n-setups/scaling-multi-main/setup.mjs
