+++ packages/@n8n/benchmark/scripts/n8n-setups/scaling-single-main/docker-compose.yml
