+++ packages/@n8n/benchmark/scripts/n8n-setups/scaling-single-main/setup.mjs
