+++ packages/@n8n/benchmark/scripts/n8n-setups/sqlite-legacy/docker-compose.yml
