+++ packages/@n8n/benchmark/scripts/n8n-setups/sqlite/docker-compose.yml
