+++ packages/@n8n/benchmark/src/n8n-api-client/authenticated-n8n-api-client.ts
@@ -1,5 +1,5 @@
 import type { AxiosRequestConfig } from 'axios';
-import { N8nApiClient } from './n8nApiClient';
+import { N8nApiClient } from './n8n-api-client';
 
 export class AuthenticatedN8nApiClient extends N8nApiClient {
 	constructor(
