+++ packages/@n8n/benchmark/src/scenario/scenario-data-loader.ts
@@ -1,7 +1,7 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import type { Scenario } from '@/types/scenario';
-import type { Workflow } from '@/n8nApiClient/n8nApiClient.types';
+import type { Workflow } from '@/n8n-api-client/n8n-api-client.types';
 
 /**
  * Loads scenario data files from FS
