+++ packages/@n8n/benchmark/src/test-execution/k6-summary.ts
